правит ошибки кода
